adaptix/services/auth/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
psycopg2-binary
django-cors-headers
celery[redis]>=5.2
kombu
requests
python-dotenv
pycryptodome
dj-database-url>=2.0.0
celery
redis
gunicorn
cryptography>=39.0.0
pyjwt[crypto]>=2.6.0
jwcrypto
django-extensions
whitenoise
drf-spectacular
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-django
opentelemetry-instrumentation-psycopg2
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp
pytest==8.0.0
pytest-django==4.8.0
pytest-mock==3.12.0
django-filter>=23.5
